Move DESI spectra viewer to repository root
